A workaround to a host_page_size() returning wrong value on iPhone 5S (16384), while host_statistics64() provides statistics relative to 4096 bytes page size.

+++ ActivityMonitor++/RAMInfo.h
@@ -12,6 +12,6 @@
 #import <Foundation/Foundation.h>
 
 @interface RAMInfo : NSObject
-@property (assign, nonatomic) NSUInteger    totalRam;
-@property (strong, nonatomic) NSString      *ramType;
+@property (assign, nonatomic) uint64_t  totalRam;
+@property (strong, nonatomic) NSString  *ramType;
 @end
